interactly_task/backend/requirements.txt
faiss-cpu
langchain
langchain-community
sentence_transformers
pandas
numpy
scikit-learn
openpyxl
flask
axios
